Add shared interface for PDF generation
