+++ apps/backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/src/helpers/pdf/PdfGeneratorHelper.ts
@@ -5,25 +5,9 @@
 // html2pdf.js only works in the browser
 
 import { PuppeteerGenerator } from './PuppeteerGenerator';
+import { HtmlPdfGeneratorInterface, PdfGeneratorOptions, RequestOptions } from './PdfGeneratorInterfaces';
 
-export type PdfGeneratorOptions = {
-  format?: 'A3' | 'A4' | 'A5' | 'Legal' | 'Letter' | 'Tabloid';
-  landscape?: boolean;
-  printBackground?: boolean;
-  margin?: {
-    top?: string;
-    bottom?: string;
-    left?: string;
-    right?: string;
-  };
-};
-
-export type RequestOptions = {
-  bearerToken?: string | null | undefined;
-  mockImageResolution?: boolean; // if true, images are mocked with a placeholder image
-};
-
-export class PdfGeneratorHelper {
+export class PdfGeneratorHelper implements HtmlPdfGeneratorInterface {
   /** Returns the default PDF generation options */
   public static getDefaultPdfGeneratorOptions(): PdfGeneratorOptions {
     return {
@@ -40,8 +24,20 @@ export class PdfGeneratorHelper {
   }
 
   /** Generates a PDF from the provided HTML string */
-  public static async generatePdfFromHtml(html: string, requestOptions: RequestOptions, options?: PdfGeneratorOptions): Promise<Buffer> {
+  public static async generatePdfFromHtml(
+    html: string,
+    requestOptions: RequestOptions,
+    options?: PdfGeneratorOptions
+  ): Promise<Buffer> {
     options = { ...this.getDefaultPdfGeneratorOptions(), ...options };
     return await PuppeteerGenerator.generatePdfFromHtmlPuppeteer(html, requestOptions, options);
+  }
+
+  public async generatePdfFromHtml(
+    html: string,
+    requestOptions: RequestOptions,
+    options?: PdfGeneratorOptions
+  ): Promise<Buffer> {
+    return await PdfGeneratorHelper.generatePdfFromHtml(html, requestOptions, options);
   }
 }
